requirements.txt
Flask 
Werkzeug 
dnspython